style: fixed time input bug on mobile screens
